feat core: reduce recv count in http processing
There is a loop in connection, which basically does `(while(recv) HttpParse();)`

Thing is, since request are almost never pipelined, this loop almost always hits ``EWOULDBLOCK`` at **every** request (for simplicity lets assume requests are <32Kb), because there is no data to read from connection at least before previous response is sent.

Say we have an established idle connection, which gets a request. Then what happens is:

1. peer_socket_.RecvSome wakes up and reads some data (lets assume the whole request): `recv` syscall
2. next recv call gets `EWOULDBLOCK` and `RecvSome` returns: one more `recv` syscall
3. Some logic for http handling gets executed
4. At the very next iteration of the loop peer_socket.RecvSome gets `EWOULDBLOCK` and goes to sleep: one more recv
5. Repeat from p.1

This is 3 recv syscalls per request, and we can reduce that number to 2 by waiting for the socket to become readable instead of unconditionally RecvSome-ing from it (and that would be done anyway after we hit initial EWOULDBLOCK).

strace for ~150k requests at current develop:
```
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ----------------
 40.28   13.391309          57    232365           epoll_wait
 17.45    5.802178          12    449698    299788 recvfrom
 12.22    4.064064          27    149892           writev
 11.50    3.824339          14    270878     28837 epoll_ctl
  7.89    2.624587          29     87901      6092 futex
  5.06    1.683114          16    105028           write
  4.64    1.541335          14    105080           read
```
strace for ~150k requests with the patch:
```
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ----------------
 28.40   10.940141          57    191749           epoll_wait
  9.95    3.832960          25    149892           writev
  9.34    3.597122          13    259782     35834 epoll_ctl
  9.32    3.590741          11    299806    149895 recvfrom
  8.18    3.150730          28    109897      7000 futex
  2.14    0.826321          14     56144           write
  2.03    0.780949          13     56169           read
```
It can be seen, that overall recv calls is 2/3 of the original value and errors count halved, which proves the idea

Tests: протестировано CI

Pull Request resolved: #151