feat logging: SetDefaultLogger and LOG_* now do not manage lifetime
Replace `logging::SetDefaultLogger` with `logging::DefaultLoggerGuard` that is only usable near `main`, outside of (wrapping) the engine lifetime. It allows to optimize logging, because it avoids accessing an `rcu::Variable` and copying a `std::shared_ptr` on the hot path of logging.

The default logger is set to the Null logger (dropping all logs) by default for non-component system applications (previously, logging to `stderr` was the default).

Tests: протестировано CI