feat core: add phdr_cache to avoid locking in unwinding
Relates: https://st.yandex-team.ru/

Multiple threads on multiple cores should be able to concurrently throw exceptions without bothering one another. But unfortunately, it appears that in the current implementation of libstdc++ and/or glibc, the stack unwinding process takes a global lock (while getting the list of shared objects, and perhaps other things) which serializes these parallel exception-throwing and can dramatically slow down the program.

Some might dismiss this inefficiency with the standard "exceptions should be rare" excuse. They should be rare. But sometimes they are not, leading do a catastrophic collapse in performance.

We saw an illustrative example of an "exception storm" in an application of ours. This application can handle lots and lots of requests per second on many cores. Some unexpected circumstance caused the application to slow down somewhat, which led to some of the requests timing out. The timeout was implemented as an exception, so now we had thousands of exceptions being thrown in all cores in parallel. This led to the applications threads starting to hang, once in a while, on the lock(s) inside "throw". This in turn made the application even slower, and created even more timeouts, which in turn resulted in even more exceptions. In this way the number of exceptions per second escalated, until the point where most of the work the application was doing was fighting over the "throw" locks, and no useful work was being done.

This patch eliminates the "throw" lock by supplying our own "dl_iterate_phdr" function which operates over a cached list of shared objects, which should mitigate blocking behavior in exception storm scenario, but as a tradeoff disables dynamic loading/unloading during component system lifetime -- there is no thread-safe and robust way to synchronize that with the cache we've got.
If one really needs dlopen/dlclose outside of component constructor/destructor, this optimization can be disabled via USERVER_DISABLE_PHDR_CACHE cmake option.

In the benchmarks which just throw+catch in parallel we are expectedly seeing X (number of threads) improvements: it tooks some 20s+ seconds for 8 threads to throw a million of exceptions each in parallel, and now it takes mere ~2s. This also improves an RPS by a factor of 2+ when an endpoint under load just throws std::runtime_error.

Some references:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=71744
https://github.com/scylladb/seastar/commit/464f5e3ae43b366b05573018fc46321863bf2fae
https://github.com/scylladb/seastar/issues/73
https://stackoverflow.com/questions/26257343/does-stack-unwinding-really-require-locks